Made a dummy refresh state in order to actually get the state to refresh on the shopping cart page
