+++ src/components/ShoppingCart.js
@@ -1,11 +1,12 @@
-import React, { useContext } from "react";
+import React, { useContext, useState } from "react";
 import { CartContext } from "../contexts/CartContext";
 
 // Components
 import Item from "./ShoppingCartItem";
 
 const ShoppingCart = () => {
-  const { cart, removeItem } = useContext(CartContext);
+  const { cart, setCart } = useContext(CartContext);
+  let [refresh, setRefresh] = useState(0);
 
   const getCartTotal = () => {
     return cart
@@ -15,10 +16,24 @@ const ShoppingCart = () => {
       .toFixed(2);
   };
 
+  function removeItem(itemId) {
+    let itemToRemove = 0;
+
+    for (let i = 0; i < cart.length; i++) {
+      if (itemId === cart[i].id) {
+        itemToRemove = cart.indexOf(cart[i]);
+      }
+    }
+    cart.splice(itemToRemove, 1);
+    let newCart = cart;
+    setCart(newCart);
+    setRefresh((refresh += 1));
+  }
+
   return (
     <div className="shopping-cart">
       {cart.map(item => (
-        <Item key={item.id} {...item} />
+        <Item key={item.id} {...item} removeItem={removeItem} />
       ))}
 
       <div className="shopping-cart__checkout">
