+++ src/components/ShoppingCartItem.js
@@ -2,7 +2,7 @@ import React, { useContext } from "react";
 import { CartContext } from "../contexts/CartContext";
 
 const Item = props => {
-  const { cart, removeItem } = useContext(CartContext);
+  //   const { removeItem } = useContext(CartContext);
 
   return (
     <div className="shopping-cart_item">
@@ -11,7 +11,9 @@ const Item = props => {
       <div>
         <h1>{props.title}</h1>
         <p>$ {props.price}</p>
-        <button onClick={() => removeItem(props.id)}>Remove from cart</button>
+        <button onClick={() => props.removeItem(props.id)}>
+          Remove from cart
+        </button>
       </div>
     </div>
   );
